requirements.txt
langchain==0.2.2
langchain_core==0.2.4
langchain_openai==0.1.8
pydantic==2.6.4
anthropic==0.35.0
